Added test to validate stock reply
